Rename Ajv Compiler Test
