+++ test/runtime/compiler-ajv/array.ts
