+++ test/runtime/compiler-ajv/boolean.ts
