+++ test/runtime/compiler-ajv/composite.ts
