+++ test/runtime/compiler-ajv/enum.ts
