+++ test/runtime/compiler-ajv/integer.ts
