+++ test/runtime/compiler-ajv/intersect.ts
