+++ test/runtime/compiler-ajv/keyof.ts
