+++ test/runtime/compiler-ajv/literal.ts
