+++ test/runtime/compiler-ajv/not.ts
