+++ test/runtime/compiler-ajv/number.ts
