+++ test/runtime/compiler-ajv/object.ts
