+++ test/runtime/compiler-ajv/omit.ts
