+++ test/runtime/compiler-ajv/optional.ts
