+++ test/runtime/compiler-ajv/partial.ts
