+++ test/runtime/compiler-ajv/pick.ts
