+++ test/runtime/compiler-ajv/readonly-optional.ts
