+++ test/runtime/compiler-ajv/readonly.ts
