+++ test/runtime/compiler-ajv/record.ts
