+++ test/runtime/compiler-ajv/recursive.ts
