+++ test/runtime/compiler-ajv/string.ts
