+++ test/runtime/compiler-ajv/void.ts
